# swagger.yaml
swagger: "2.0"
info:
  description: "This is the swagger doc for vanDevTimes"
  version: "1.0.0"
  title: "Swagger vanDevTimes"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "vantimes.info@gmail.com"
basePath: "/api"
paths:
...
  /categories:
    get:
      tags:
        - "Category"
      summary: "Get all content categories"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Category"
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
    post:
      tags:
        - "Category"
      summary: "add category with name"
      description: ""
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "category name"
          required: true
          schema:
            $ref: "#/definitions/Category"
      responses:
        200:
          description: "successful operation"
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
    delete:
      tags:
        - "Category"
      summary: "remove categroy with name after re-categorizing all content in name to Uncategorized"
      description: ""
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "category name"
          required: true
          schema:
            example: "testCategory"
      responses:
        200:
          description: "successful operation"
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
...
definitions:
  Category:
    type: "object"
    properties:
      categoryName:
        type: "string"
...